+++ themes-default/slim/src/components/display-show.vue
@@ -110,7 +110,12 @@
                         <span v-else-if="props.column.label == 'Subtitles'" class="align-center">
                             <div class="subtitles">
                                 <div v-for="flag in props.row.subtitles" :key="flag">
-                                    <img v-if="flag !== 'und'" :src="`images/subtitles/flags/${flag}.png`" width="16" height="11" :alt="flag" onError="this.onerror=null;this.src='images/flags/unknown.png';" @click="searchSubtitle($event, props.row, flag)">
+                                    <img
+                                        v-if="flag !== 'und'" :src="`images/subtitles/flags/${flag}.png`"
+                                        width="16" height="11" :alt="flag" class="addQTip" :title="flag"
+                                        onError="this.onerror=null;this.src='images/flags/unknown.png';"
+                                        @click="searchSubtitle($event, props.row, flag)"
+                                    >
                                     <img v-else :src="`images/subtitles/flags/${flag}.png`" class="subtitle-flag" width="16" height="11" :alt="flag" onError="this.onerror=null;this.src='images/flags/unknown.png';">
                                 </div>
                             </div>
@@ -257,7 +262,12 @@
                         <span v-else-if="props.column.label == 'Subtitles'" class="align-center">
                             <div class="subtitles">
                                 <div v-for="flag in props.row.subtitles" :key="flag">
-                                    <img v-if="flag !== 'und'" :src="`images/subtitles/flags/${flag}.png`" width="16" height="11" alt="{flag}" onError="this.onerror=null;this.src='images/flags/unknown.png';" @click="searchSubtitle($event, props.row, flag)">
+                                    <img
+                                        v-if="flag !== 'und'" :src="`images/subtitles/flags/${flag}.png`"
+                                        width="16" height="11" :alt="flag" class="addQTip" :title="flag"
+                                        onError="this.onerror=null;this.src='images/flags/unknown.png';"
+                                        @click="searchSubtitle($event, props.row, flag)"
+                                    >
                                     <img v-else :src="`images/subtitles/flags/${flag}.png`" class="subtitle-flag" width="16" height="11" alt="flag" onError="this.onerror=null;this.src='images/flags/unknown.png';">
                                 </div>
                             </div>