+++ CHANGELOG.md
@@ -8,7 +8,8 @@
 
 #### Fixes
 - Homepage: Fix loading shows from localCache ([10779](https://github.com/pymedusa/Medusa/pull/10779))
-
+- Fix Erai-raws formatted anime release guessit parsing ([10779](https://github.com/pymedusa/Medusa/pull/10779))
+ 
 -----
 
 ## 1.0.3 (10-06-2022)